add navbar

#17 add navbar
