+++ functions/assets/navbar.css
@@ -1,0 +1,71 @@
+#navContainer {
+    background-color: #f2f2f2;
+    border: 1px solid #ccc;
+    padding: 0;
+    display: flex;
+    justify-content: space-between;
+    margin: 0 0 1em 0;
+}
+
+#leftHeaderContainer {
+    display: flex;
+    justify-content: flex-start;
+}
+
+#nav ul {
+    list-style: none;
+    margin: 0 auto;
+    padding: 0;
+}
+
+#nav li {
+    float: left;
+}
+
+#nav li a {
+    display: block;
+    padding: 8px 15px;
+    text-decoration: none;
+    font-weight: bold;
+    color: #069;
+    border-right: 1px solid #ccc;
+}
+
+#nav li:first-child a {
+    border-left: 1px solid #ccc;
+}
+
+#nav li a:hover {
+    color: #c00;
+    background-color: #fff;
+}
+
+#middleHeaderContainer {
+    display: flex;
+    justify-content: center;
+    padding-right: 10%;
+}
+
+#headerTitle {
+    color: #333333;
+    font-family: 'Bitter', serif;
+    font-size: x-large;
+    font-weight: normal;
+}
+
+#logoutButtonBox {
+    margin: 0 auto;
+    padding: 0;
+}
+
+#logoutButton {
+    display: block;
+    padding: 8px 15px;
+    text-decoration: none;
+    font-weight: bold;
+    color: #069;
+}
+
+#logoutButton:hover {
+    cursor: pointer;
+}
