+++ functions/controller/application.controller.js
@@ -1,5 +1,6 @@
 const Availability = require('../model/availability.model')
 const Sequelize = require('sequelize')
+const CompetenceProfile = require('../model/competence_profile.model')
 
 /**
  * To check the availability if the person have this time
@@ -29,4 +30,33 @@ const registerAvailability = async (person_id, from_date, to_date) => {
   }
 }
 
-module.exports = { registerAvailability }
+const registerCompetence = async (
+  person_id,
+  competence_id,
+  years_of_experience,
+) => {
+  try {
+    const competenceProfileExists = await CompetenceProfile.findOne({
+      where: {
+        [Sequelize.Op.and]: [
+          { person_id },
+          { competence_id },
+          { years_of_experience },
+        ],
+      },
+    })
+    if (competenceProfileExists) {
+      throw new Error('Competence profile already exists')
+    }
+    const newCompetenceProfile = await CompetenceProfile.create({
+      person_id,
+      competence_id,
+      years_of_experience,
+    })
+    return newCompetenceProfile
+  } catch (error) {
+    throw error
+  }
+}
+
+module.exports = { registerAvailability, registerCompetence }
