+++ functions/views/application-form.ejs
@@ -10,6 +10,7 @@
 </head>
 
 <body>
+  <%- include('./partials/navbar.ejs') %> 
   <form method="post">
 
     <!-- Error for from_date -->
