+++ functions/views/applications.ejs
@@ -8,6 +8,7 @@
 </head>
 
 <body>
+    <%- include('./partials/navbar.ejs') %>
     HELLO
 </body>
