+++ functions/views/home.ejs
@@ -6,6 +6,7 @@
     <title>Document</title>
 </head>
 <body>
+    <%- include('./partials/navbar.ejs') %>
     <h1>SUCCESSS</h1>
     <h1><%= user.username %></h1>
