+++ functions/views/login.ejs
@@ -7,6 +7,7 @@
         <title>Login</title>
     </head>
     <body>
+        <%- include('./partials/navbar.ejs') %>
         <form id="signInBox" method="post">
             <div id="signin">
                 <h2>Sign In</h2>
