+++ functions/views/partials/navbar.ejs
@@ -1,0 +1,23 @@
+<link href="../assets/navbar.css" rel="stylesheet" type="text/css">
+<div id="navContainer">
+    <div id="leftHeaderContainer">
+        <nav id="nav">
+            <ul id="navOptions">
+                <li><a href="/">Home</a></li>
+                <li><a href="/about">About</a></li>
+                <li><a href="/posts/forumpage">Posts</a></li>
+                <li><a href="/account">Account</a></li>
+            </ul>
+        </nav>
+    </div>
+    <div id="middleHeaderContainer">
+        <span id="headerTitle">Recruitment Application</span>
+    </div>
+    <div id="leftHeaderContainer">
+        <form id="logoutButtonBox" action="/auth/logout">
+            <button id="logoutButton" type="submit">Sign Out</button>
+        </form>
+    </div>
+</div>
+
+
