+++ functions/views/register.ejs
@@ -7,6 +7,7 @@
         <title>Register</title>
     </head>
     <body>
+        <%- include('./partials/navbar.ejs') %>
         <form id="signInBox" method="post">
             <p style="font-size: xx-small; color: red;" ><%= error %></p>
             <div id="signin">
